Corrections
Multiple corrections to how the names of images were written. If we follow the w3 schools standards ALL images should be written in LOWER CASE letters. This prevented the images from rendering. There are 2 index pages in the web the only reason your site is rendering is because of the captialization of one of the index pages. I also corrected you javascript error.